+++ Core/Entity/Projectile/ProjectilePee.cs
@@ -7,7 +7,7 @@ namespace TBoGV;
 internal class ProjectilePee : Projectile
 {
 	public static Texture2D Sprite { get; protected set; }
-	public ProjectilePee(Vector2 position, Vector2 direction, int damage)
+	public ProjectilePee(Vector2 position, Vector2 direction, float damage)
 	{
 		Sprite = TextureManager.GetTexture("projectile");
 		// Size = new Vector2(7, 7);
