+++ Core/Entity/Projectile/ProjectileRoot.cs
@@ -7,7 +7,7 @@ namespace TBoGV;
 internal class ProjectileRoot : Projectile
 {
 	public static Texture2D Sprite { get; protected set; }
-	public ProjectileRoot(Vector2 position, Vector2 direction, int damage) 
+	public ProjectileRoot(Vector2 position, Vector2 direction, float damage) 
 	{
 		Sprite = TextureManager.GetTexture("koren");
 		// Size = new Vector2(7, 7);
