+++ Core/Interface/IRecieveDmg.cs
@@ -1,7 +1,7 @@
 ﻿namespace TBoGV;
 internal interface IRecieveDmg
 {
-	int Hp { get; set; }
+	float Hp { get; set; }
 	int MaxHp { get; set; }
-	void RecieveDmg(int damage);
+	void RecieveDmg(float damage);
 }
